Trigger Maven actions in appropriate subprojects
If the main project doesn't have an execution defined for the requested
goal, trigger the action in all subprojects. If none of them has a
corresponding execution, fall back to running the action without checks
in the selected project.

Workspace artifact resolution is now enabled when running from Vaadin
toolbar.

Closes #670