Extend m2e build for themes and update-widgetset
The build participant executes the update-theme and compile-theme
mojos only if there are theme directories with modifications, and
refreshes theme directories after the build.

It also runs update-theme as a part of the build only when there
are related changes.

All these are conditional to having the corresponding executions
defined in the POM.

Closes #671